+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Add basic mobile support
   - Add custom MQM taxonomy support
   - Add `data_welcome` to support tutorials seen by all annotators
+  - Improve user progress reporting in dashboard
+  - Add `docs_per_user` cap for single-stream and dynamic assignment
 - v1.0.2
   - Fix `token_main` vulnerability
   - Add textfields for translation and post-editing